Update test to handle replica indexes and deleting duplicates
